pyproject.toml
[project]
name = "face-comm"
version = "0.1.0"
description = "顔ジェスチャーコミュニケーションシステム - 顔の動きでコミュニケーションを支援"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0,<0.10.15",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
gui = [
    "pyqt5>=5.15.0",
]
audio = [
    "pyttsx3>=2.90",
]

[project.scripts]
face-comm = "src.main:main"
face-comm-debug = "src.gesture_detector:main"
